pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool]
[tool.poetry]
authors = ["takipsizad <addeniz25@gmail.com>"]
description = "my bot"
license = "GNU General Public License v3.0"
name = "bot"
version = "1.0"
[tool.poetry.dependencies]
Flask = "^1.1.2"
aiodns = "^2.0.0"
bs4 = "^0.0.1"
dblpy = "^0.4.0"
discord-py-slash-command = "^1.1.0"
"discord.py" = "^1.6.0"
jishaku = "^1.20.0"
motor = "^2.3.1"
pycoingecko = "^1.4.1"
python = "^3.8"
pytube = "^10.6.1"
tornado = "^6.1"
python-dotenv = "^0.17.1"
psutil = "^5.8.0"
dnspython = "^2.1.0"
dpymenus = "^2.1.5"
[tool.poetry.dev-dependencies]
